% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalling.R
\name{scale_by_median}
\alias{scale_by_median}
\title{scaling}
\usage{
scale_by_median(df, category)
}
\arguments{
\item{df}{Dataframe(I use df as the abbreviation of data frame)}

\item{category}{Categorial variable (based on which I cauculate the median for each group)}
}
\value{
the funtion returns the scaled dataset based on input \code{df} and \code{category}
}
\description{
This function scales all of its numeric variables. Specifically, I scale each observation the median number in the category.
}
\examples{
scale_by_median(mtcars,mtcars$cyl)
}
